Fixed (unsigned long) and (long) castings to avoid compiler warnings.
Fixed bug in TKImageExportController which retained an NSArray rather than initializing the presets mutable array by using -mutableCopy.

git-svn-id: file:///Users/mdouma46/Developer/sourceFinagler/branches/2.0.3@118 c3d3028b-b5cd-4c35-8691-6fb8b2347b27

+++ SourceFinagler/MDAppKitAdditions.m
@@ -347,7 +347,7 @@ NSString *NSStringFromDefaultsKeyPath(NSString *defaultsKey) {
 #endif
 	NSRect frameRect = [self frame];
 	
-	NSString *dimensionString = [NSString stringWithFormat:@"%ld %ld %ld %ld", (NSInteger)frameRect.origin.x, (NSInteger)frameRect.origin.y, (NSInteger)frameRect.size.width, (NSInteger)frameRect.size.height];
+	NSString *dimensionString = [NSString stringWithFormat:@"%ld %ld %ld %ld", (long)frameRect.origin.x, (long)frameRect.origin.y, (long)frameRect.size.width, (long)frameRect.size.height];
 	
 	return dimensionString;
 }
